feat: CategoryServiceImpl (admin) is ready

+++ main_svc/src/main/java/ru/practicum/ewm/main_svc/service/impl/CategoryServiceImpl.java
@@ -1,7 +1,10 @@
 package ru.practicum.ewm.main_svc.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.ewm.main_svc.error.EwmException;
 import ru.practicum.ewm.main_svc.model.dto.CategoryDto;
 import ru.practicum.ewm.main_svc.model.dto.NewCategoryDto;
 import ru.practicum.ewm.main_svc.model.util.mappers.CategoryMapper;
@@ -15,28 +18,40 @@ public class CategoryServiceImpl implements CategoryService {
     private final CategoryMapper categoryMapper;
 
     @Override
+    @Transactional
     public CategoryDto adminCreate(NewCategoryDto newCategoryDto) {
         return categoryMapper.entity2categoryDto(categoryRepository.save(categoryMapper.categoryDto2entity(newCategoryDto)));
     }
 
     @Override
+    @Transactional
     public void adminDelete(Long catId) {
-        categoryRepository.deleteById(catId);
+        if (categoryRepository.existsById(catId)) categoryRepository.deleteById(catId);
+        else throw new EwmException(String.format("Category with id=%d was not found.", catId), HttpStatus.NOT_FOUND);
     }
 
     @Override
+    @Transactional
     public CategoryDto adminUpdate(Long catId, CategoryDto categoryDto) {
-        var category = categoryMapper.categoryDto2entity(categoryDto);
-        category.setId(catId);
-        return categoryMapper.entity2categoryDto(categoryRepository.save(category));
+        var category = categoryRepository.findById(catId)
+                .orElseThrow(() -> new EwmException(String.format("Category with id=%d was not found.", catId), HttpStatus.NOT_FOUND));
+
+        if (!categoryDto.getName().equals(category.getName())) {
+            category.setName(categoryDto.getName());
+            category = categoryRepository.save(category);
+        }
+
+        return categoryMapper.entity2categoryDto(category);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Iterable<CategoryDto> publicFindAll(Integer from, Integer size) {
         return null;
     }
 
     @Override
+    @Transactional(readOnly = true)
     public CategoryDto publicFindById(Long catId) {
         return null;
     }
